+++ tests/_async/test_transport.py
@@ -122,3 +122,58 @@ async def test_transport_stale_response_with_connecterror(use_temp_dir):
             await cache_transport.handle_async_request(request)
             response = await cache_transport.handle_async_request(request)
             assert response.extensions["from_cache"]
+
+
+@pytest.mark.anyio
+async def test_transport_with_only_if_cached_directive_without_stored_response(
+    use_temp_dir,
+):
+    controller = hishel.Controller()
+
+    async with hishel.MockAsyncTransport() as transport:
+        async with hishel.AsyncCacheTransport(
+            transport=transport, controller=controller
+        ) as cache_transport:
+            response = await cache_transport.handle_async_request(
+                httpx.Request(
+                    "GET",
+                    "https://www.example.com",
+                    headers=[(b"Cache-Control", b"only-if-cached")],
+                )
+            )
+            assert response.status_code == 504
+
+
+@pytest.mark.anyio
+async def test_transport_with_only_if_cached_directive_with_stored_response(
+    use_temp_dir,
+):
+    controller = hishel.Controller()
+
+    async with hishel.MockAsyncTransport() as transport:
+        transport.add_responses(
+            [
+                httpx.Response(
+                    200,
+                    headers=[
+                        (b"Cache-Control", b"max-age=3600"),
+                        (b"Date", b"Mon, 25 Aug 2015 12:00:00 GMT"),
+                    ],
+                    content=b"test",
+                ),
+            ]
+        )
+        async with hishel.AsyncCacheTransport(
+            transport=transport, controller=controller
+        ) as cache_transport:
+            await cache_transport.handle_async_request(
+                httpx.Request("GET", "https://www.example.com")
+            )
+            response = await cache_transport.handle_async_request(
+                httpx.Request(
+                    "GET",
+                    "https://www.example.com",
+                    headers=[(b"Cache-Control", b"only-if-cached")],
+                )
+            )
+            assert response.status_code == 504